DMD-717 Instant Send Fee Changes (Implement UI)
Added support for InstantSend Auto Locks.
This commit removes the "Use InstantSend" switch from Send Dash screen. The new flow works as follows:
If IS and Auto Locks are supported by the network, the wallet will try to send all the transactions using IS Auto Lock, if the transaction doesn't meet the conditions of being a simple transaction and can't be auto locked, but user has enough funds to send it as a standard IS, she will be asked if wants to add additional fee and force IS or send it as a regular transaction, if she doesn't have enough funds to send it as a standard IS it will be automatically send as regular transaction.