ScanActivity: Fix IllegalStateException on Android 6.0 and below (#174)
ScanActivity: Work around IllegalStateException in 
FragmentManagerImpl.checkStateLoss() on Android 6.0 and below.

Apply changes from Bitcoin Wallet with a small delay in 
onRequestPermissionsResult.  Without it a crash results from time
to time.

Replace ?dialogPreferredPadding in resources with hardcoded values 
of 24dp.  The dialogPreferredPadding attribute seems to cause crash 
on some devices.